timer: Transform timer into a systemd service
The extension must not delay disable. #10
However, the first time the unlock-dialog is enabled after login, GNOME shell
quickly disables and enables all extensions.
In order to allow the check-command to run w/o interruption while fully disabled,
the `Timer` class is run in a separate gjs process as a DBus service.